delete works now but need to ensure unique images delete from storage and firestore

+++ lib/services/message_card_delete_service.dart
@@ -7,14 +7,33 @@ class MessageCardDeleteService {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
   Future<void> deleteMessageCard(String profileId, String cardId) async {
+    print('message card delete service called for profileId $profileId cardId $cardId');
     try{
-      await _firestore
+      //first find the document with the given MessageCardId
+      QuerySnapshot querySnapshot = await _firestore
           .collection('profiles')
           .doc(profileId)
           .collection('messageCards')
-          .doc(cardId)
-          .delete();
-      print('Message card $cardId deleted successfully');
+          .where('messageCardId', isEqualTo: cardId)
+          .limit(1)
+          .get();
+
+      if(querySnapshot.docs.isNotEmpty) {
+        String docId  = querySnapshot.docs.first.id;
+
+        //now delete the doc
+        await _firestore
+        .collection('profiles')
+        .doc(profileId)
+        .collection('messageCards')
+        .doc(docId)
+        .delete();
+
+        print('Message card $cardId deleted successfully');
+      }else{
+        print('Message card $cardId not found');
+      }
+
     }catch (e) {
       print('Failed to delete message card $cardId: $e');
       rethrow;
